requirements.txt
django
djangorestframework
django-cors-headers
channels_redis
channels
APScheduler
pyserial
